Pin the encrypted-model hooks fix with a compile-pass case
Adds `repository_encrypted_hooks.rs` to the trybuild compile-pass set so
the `AsChangeset` regression is caught at the macro level, independent of
the CLI generator that made it reachable.

Uses an explicit `hooks = T` rather than `broadcasts = true`: both reach
the same hooks-aware `update_many` path, but `broadcasts` pulls in
`autumn_web::channels`, which is feature-gated — the trybuild harness
runs at default features, so the broadcasts spelling would have failed
for an unrelated reason.
